Create README.md
The MySQL Proxy Gateway is a TCP proxy that facilitates communication between Docker containers and a MySQL server. It listens on a specified proxy port, forwards the data between Docker clients and MySQL servers, and manages both incoming and outgoing data streams. This system is typically used to proxy MySQL queries from Docker containers to a MySQL server running locally or on a remote host.